+++ clash-lib/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for the [`clash-lib`](http://hackage.haskell.org/package/clash-lib) package
 
+## 0.7
+* New features:
+  * Support for `clash-prelude` 0.11
+
 ## 0.6.21
 * Fixes bugs:
   * Type families are not currently being reduced correctly [#167](https://github.com/clash-lang/clash-compiler/issues/167)